Start porting to pytest
Instead of using a custom test runner, we will try to run all tests with pytest.
This is a first step on making it possible to run at least a subset of the tests as part of a CI setup.

For that the xstrct_run.py file had to be reorganized.
A main function (in a separate file) has been introduced which can be called from the tests and runs a simulation.