Use only one ofstream per custom output file
Previously the ofstream's were created for every write access
(e.g. every call of syn_scale). In small benchmarks that compared
opening the files for every write access and opening the files once
and then writing just as often the difference in runtime was
>1min vs 1.5seconds.
The same benchmarks couldn't be tested on the cluster, because the
program with reopening the files for every write ran for >20minutes.
The simulation is not writing to the files as often, but considering
NFS performance this might still improve simulation time.