Use only one ostream per output file
This time take into consideration that the C++ functions in cpp_methods.py
will be defined in multiple translation units but all definitions should
share the same ofstreams. We achieve this by creating a new translation unit
which holds the ofstreams, and which is being linked to by all other units.