Introduce strict structural homeostasis
The previous homeostasis mechanism would simply increase/decrease the insertion probability of new synapses.
This new mechanism adjusts the growth rate exactly such that the missing number of synapes is inserted.
Due to how structural plasticity works this is a stochastic process, but tests indicate the with the new mechanism the number of active synapses stays reasonably constant.

It is compatible with distance dependent connections.